+++ src/coinbase/types.ts
@@ -27,6 +27,7 @@ import {
   StakingContext as StakingContextModel,
   FetchStakingRewardsRequest,
   FetchStakingRewards200Response,
+  FetchHistoricalStakingBalances200Response,
   FaucetTransaction,
   BroadcastStakingOperationRequest,
   CreateStakingOperationRequest,
@@ -441,6 +442,29 @@ export type StakeAPIClient = {
     page?: string,
     options?: AxiosRequestConfig,
   ): AxiosPromise<FetchStakingRewards200Response>;
+
+  /**
+   * Get the staking balances for an address.
+   *
+   * @param address - The onchain address to fetch the staking balances for.
+   * @param networkId - The ID of the blockchain network.
+   * @param assetId - The ID of the asset to fetch the staking balances for.
+   * @param startTime - The start time of the staking balances.
+   * @param endTime - The end time of the staking balances.
+   * @param limit - The amount of records to return in a single call.
+   * @param page - The batch of records for a given section in the response.
+   * @param options - Axios request options.
+   */
+  fetchHistoricalStakingBalances(
+    address: string,
+    networkId: string,
+    assetId: string,
+    startTime: string,
+    endTime: string,
+    limit?: number,
+    page?: string,
+    options?: AxiosRequestConfig,
+  ): AxiosPromise<FetchHistoricalStakingBalances200Response>;
 
   broadcastStakingOperation(
     walletId: string,