+++ src/tests/external_address_test.ts
@@ -16,6 +16,7 @@ import {
   AddressBalanceList,
   Balance,
   FetchStakingRewards200Response,
+  FetchHistoricalStakingBalances200Response,
   StakingContext as StakingContextModel,
   StakingOperation as StakingOperationModel,
   StakingRewardFormat,
@@ -28,6 +29,7 @@ import { StakingOperation } from "../coinbase/staking_operation";
 import { Asset } from "../coinbase/asset";
 import { randomUUID } from "crypto";
 import { StakingReward } from "../coinbase/staking_reward";
+import { StakingBalance } from "../coinbase/staking_balance";
 
 describe("ExternalAddress", () => {
   const newAddress = newAddressModel("", randomUUID(), Coinbase.networks.EthereumHolesky);
@@ -128,6 +130,55 @@ describe("ExternalAddress", () => {
           conversion_price: "3000",
           conversion_time: "2024-05-03T00:00:00Z",
         },
+      },
+    ],
+    has_more: false,
+    next_page: "",
+  };
+
+  const HISTORICAL_STAKING_BALANCES_RESPONSE: FetchHistoricalStakingBalances200Response = {
+    data: [
+      {
+        address: address.getId(),
+        date: "2024-05-01",
+        bonded_stake: {
+          amount: "32",
+          asset: {
+            asset_id: Coinbase.assets.Eth,
+            network_id: address.getNetworkId(),
+            decimals: 18,
+          },
+        },
+        unbonded_balance: {
+          amount: "2",
+          asset: {
+            asset_id: Coinbase.assets.Eth,
+            network_id: address.getNetworkId(),
+            decimals: 18,
+          },
+        },
+        participant_type: "validator",
+      },
+      {
+        address: address.getId(),
+        date: "2024-05-02",
+        bonded_stake: {
+          amount: "33",
+          asset: {
+            asset_id: Coinbase.assets.Eth,
+            network_id: address.getNetworkId(),
+            decimals: 18,
+          },
+        },
+        unbonded_balance: {
+          amount: "3",
+          asset: {
+            asset_id: Coinbase.assets.Eth,
+            network_id: address.getNetworkId(),
+            decimals: 18,
+          },
+        },
+        participant_type: "validator",
       },
     ],
     has_more: false,
@@ -371,6 +422,26 @@ describe("ExternalAddress", () => {
           end_time: endTime,
           format: StakingRewardFormat.Usd,
         },
+        100,
+        undefined,
+      );
+    });
+  });
+
+  describe(".historicalStakingBalances", () => {
+    it("should return staking balances successfully", async () => {
+      Coinbase.apiClients.stake!.fetchHistoricalStakingBalances = mockReturnValue(HISTORICAL_STAKING_BALANCES_RESPONSE);
+      Coinbase.apiClients.asset!.getAsset = getAssetMock();
+      const response = await address.historicalStakingBalances(Coinbase.assets.Eth, startTime, endTime);
+
+      expect(response).toBeInstanceOf(Array<StakingBalance>);
+      expect(response.length).toEqual(2);
+      expect(Coinbase.apiClients.stake!.fetchHistoricalStakingBalances).toHaveBeenCalledWith(
+        address.getId(),
+        address.getNetworkId(),
+        Coinbase.assets.Eth,
+        startTime,
+        endTime,
         100,
         undefined,
       );