+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - Add Function `listHistoricalBalances` for `Address` for fetching historical balances for an asset
+- Support for retrieving historical staking balances information
 
 ## [0.0.15]